[TRANSFORMATION-INTEG] Transformation profiles for Report Comparison configured in the settings file.
The Report Comparison application needs to be able to build transformer chain with settings from a JSON the same way as the Rewriter. The infrastructure for this will be moved to a separate project later, and therefore, in order to simply make things work, the Comparison project references the Rewriter for now.
The functionality of the Comparison tool is not affected, and therefore only the PayPal transaction filter is available.